fix: apex code coverage flag
Changed apex test run payload to set skipCodeCoverage. When code coverage is not called for,
skipCodeCoverage is now set to true. It is set to false when code coverage is specified. This fixes
an issue where code coverage was being run with every test, because this flag was omitted. Updated
corresponding tests to check for payload change.

#3541