fix: fix synchronous run error
Resolved issue where including the synchronous flag without a result format resulted in an error
message. Command now runs as it did prior to toolbelt migration, where we wait until the results are
collected and then display them in human form. Note this is not an actual synchronous run, as it's
too many tests for the API, and we have a ticket out for W-9163533 to address that.