fix: junit testresult file generation (#285)
Resolved an issue where the junit file generation created invalid XML files, because the failure message
contains characters that should be XML-escaped. These are: <, >, &, " and '.
Made sure to xml-encode these in the message, updated a testcase to make sure that this is now covered in
a test as well.

Additional commit to also escape the method name, it can be '<compile>' when testcases fail because of
compilation errors. The '<' and '>' characters also need to be escaped in that case.

Co-authored-by: Johannes Verelst <johannes@verelst.net>